[refactor] InputField props 수정
