+++ src/components/auth/shared/ui/InputField.tsx
@@ -3,10 +3,7 @@ import Image from 'next/image';
 
 interface InputFieldProps extends React.InputHTMLAttributes<HTMLInputElement> {
     label: string;
-    id: string;
-    type: string;
-    placeholder: string;
-    isSubmitted: boolean;
+    disabled: boolean;
     error?: string;
     errorResponseMessage?: string | null;
     isPasswordVisible?: boolean;
@@ -14,7 +11,7 @@ interface InputFieldProps extends React.InputHTMLAttributes<HTMLInputElement> {
 }
 
 const InputField = React.forwardRef<HTMLInputElement, InputFieldProps>(
-    ({ label, id, type, placeholder, error, errorResponseMessage, isSubmitted, isPasswordVisible, handlePasswordVisibility, ...props }, ref) => (
+    ({ label, id, type, placeholder, error, errorResponseMessage, disabled, isPasswordVisible, handlePasswordVisibility, ...props }, ref) => (
         <div className="w-full flex flex-col gap-2">
             <label htmlFor={id} className="block text-sm text-gray-900 font-bold">
                 {label}
@@ -24,9 +21,10 @@ const InputField = React.forwardRef<HTMLInputElement, InputFieldProps>(
                     ref={ref}
                     type={label === '비밀번호' ? (isPasswordVisible ? 'text' : 'password') : type}
                     id={id}
-                    className={`block w-full p-2.5 rounded-lg bg-gray-50 text-sm text-gray-900 border-2 focus:outline-none ${error || errorResponseMessage ? 'border-red-600' : 'focus:border-main-300'}`}
+                    disabled={disabled}
                     placeholder={placeholder}
-                    aria-invalid={isSubmitted ? (error ? 'true' : 'false') : undefined}
+                    aria-invalid={disabled ? (error ? 'true' : 'false') : undefined}
+                    className={`block w-full p-2.5 rounded-lg bg-gray-50 text-sm text-gray-900 border-2 focus:outline-none ${error || errorResponseMessage ? 'border-red-600' : 'focus:border-main-300'}`}
                     {...props}
                 />
                 {label === '비밀번호' && (
