+++ src/components/gatherings/detail/Footer.tsx
