+++ src/components/gatherings/detail/Information.tsx
