+++ src/components/gatherings/detail/Thumbnail.tsx
