+++ src/components/gatherings/detail/loading/InformationLoading.tsx
